backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
# ── Database ──
sqlalchemy>=2.0.0
# ── Auth ──
PyJWT>=2.8.0
python-dotenv>=1.0.0
